Update burger menu script
